requirements.txt
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.7.1
diffusers==0.31.0
distro==1.9.0
exceptiongroup==1.2.2
fairscale==0.4.4
fastapi==0.115.3
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
gradio==4.44.1
gradio_client==1.3.0
# Editable install with no version control (groundingdino==0.1.0)
-e f:\program_learn\project\grounded-segment-anything\groundingdino
h11==0.14.0
hf_transfer==0.1.8
hmmlearn==0.3.2
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
litellm==1.51.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
nltk==3.9.1
numpy==1.23.2
onnxruntime==1.19.2
openai==1.52.2
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
-e git+https://github.com/xinyu1205/recognize-anything.git@88c2b0ca13e38cca6655f83ad0185271167dbcbf#egg=ram
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
ruff==0.7.1
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.0
# Editable install with no version control (segment_anything==1.0)
-e f:\program_learn\project\grounded-segment-anything\segment_anything
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
supervision==0.24.0
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.19.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
websockets==12.0
yapf==0.40.2
yarl==1.15.2
zipp==3.20.2
